Finished Javascript
For this commit I used the HTML and CSS code from the coin flip game as a base. Then, I tweaked it to fit a rock paper scissors game. I then did all the javascript to get the game working. The next step will now be to improve the HTML and CSS codes.